+++ anonymous-board/src/test/java/com/example/anonymousboard/post/controller/PostControllerTest.java
@@ -30,7 +30,7 @@ import com.example.anonymousboard.post.exception.InvalidPostKeywordException;
 import com.example.anonymousboard.post.exception.InvalidTitleException;
 import com.example.anonymousboard.post.exception.PostErrorCode;
 import com.example.anonymousboard.post.exception.PostNotFoundException;
-import com.example.anonymousboard.util.controller.ControllerTest;
+import com.example.anonymousboard.util.ControllerTest;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
