+++ anonymous-board/src/test/java/com/example/anonymousboard/post/repository/PostRepositoryTest.java
@@ -3,7 +3,7 @@ package com.example.anonymousboard.post.repository;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.example.anonymousboard.post.domain.Post;
-import com.example.anonymousboard.util.controller.RepositoryTest;
+import com.example.anonymousboard.util.RepositoryTest;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
